requirements.txt
# requirements.txt
numpy
scipy
matplotlib
plotly
pyyaml
python-dotenv
pytest # For testing
packaging